+++ Sources/KeyboardKit/KeyboardInputViewController.swift
@@ -280,8 +280,11 @@ open class KeyboardInputViewController: UIInputViewController, KeyboardControlle
     /**
      The autocomplete provider that is used to provide users
      with autocomplete suggestions.
+     
+     > v8.0: This property will be made optional, which will
+     let us remove the `DisabledAutocompleteProvider`.
      */
-    public lazy var autocompleteProvider: AutocompleteProvider = DisabledAutocompleteProvider()
+    public lazy var autocompleteProvider: AutocompleteProvider = DeprecatedAutocompleteProvider()
 
     /**
      The callout action provider that is used to provide the