+++ Tests/KeyboardKitTests/Keyboard/KeyboardInputViewControllerTests.swift
@@ -122,7 +122,7 @@ class KeyboardInputViewControllerTests: XCTestCase {
     // MARK: - Services
 
     func servicesHaveStandardInstancesByDefault() {
-        XCTAssertNotNil(vc.autocompleteProvider as? DisabledAutocompleteProvider)
+        XCTAssertNotNil(vc.autocompleteProvider as? DeprecatedAutocompleteProvider)
         XCTAssertNotNil(vc.calloutActionProvider as? StandardCalloutActionProvider)
         XCTAssertNotNil(vc.keyboardActionHandler as? StandardKeyboardActionHandler)
         XCTAssertNotNil(vc.keyboardBehavior as? StandardKeyboardBehavior)